castd: logging tweaks
This is the first patch in the series rebuilding the cast daemon. It
changes the way logging in the daemon works, makes use of the libcast
log callback and introduces the context structure which will be used
to pass around crucial data within the daemon.

Also: remove the atomic builtins from logging.c. They do not offer
proper protection against concurrent access and the logging config
routines should not be really used that often anyway. Maybe we'll go
back to using pthread mutexes in the future.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>